+++ processing/schematisation_algorithms.py
@@ -70,9 +70,11 @@ class MigrateAlgorithm(QgsProcessingAlgorithm):
         schema = ModelSchema(threedi_db)
         try:
             schema.validate_schema()
+            schema.set_spatial_indexes()
         except errors.MigrationMissingError:
             backup_filepath = backup_sqlite(filename)
             schema.upgrade(backup=False, upgrade_spatialite_version=True)
+            schema.set_spatial_indexes()
             shutil.rmtree(os.path.dirname(backup_filepath))
         except errors.UpgradeFailedError:
             feedback.pushWarning(
@@ -149,6 +151,8 @@ class CheckSchematisationAlgorithm(QgsProcessingAlgorithm):
         threedi_db = get_threedi_database(filename=input_filename, feedback=feedback)
         if not threedi_db:
             return {self.OUTPUT: None}
+        schema = ModelSchema(threedi_db)
+        schema.set_spatial_indexes()
         try:
             model_checker = ThreediModelChecker(threedi_db)
         except errors.MigrationMissingError:
@@ -290,6 +294,7 @@ class CheckRastersAlgorithm(QgsProcessingAlgorithm):
         try:
             schema = ModelSchema(threedi_db)
             schema.validate_schema()
+            schema.set_spatial_indexes()
             checker = RasterChecker(threedi_db)
             checker.run_all_checks(feedback=feedback)
             checker.close_session()